desafio1CH/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-ckeditor==6.2.0
django-js-asset==2.0.0
Pillow==9.0.1
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2
Unipath==1.1